now routers are actually working, and I have a blank separate admin page from my zombie parade
